Custom cell-formatting support
Enhances table allowing custom formatting of cells by giving you a
callback function to modify the final cell output before rendering.

In this way, you can enjoy the formatting you selected and customize it
further, for example, consider you have to display seconds with decimal
part and want to "decorate" the number with a final "s" in order to make
it explicit. Easy, pass a function like this:

    def all_but_headers(x, y, cell):
        if y==0: # headers
            return cell
        return cell+"s"

 Everything else will work as always.

 I've chosen to make the y coordinate to mean *always* the header line
 (whenever it be present or not) and all rows will be numbered from 1
 on. Columns will be zero based.

 As I'm unsure about whether or not you want to keep it compatible with
 python 2.1, I've added a kind-of enumerate compatible function enum(..)
 which should be compatible but doesn't need generators or anything
 else.
 It can be used in the future to "simplify" expressions like this:
 `for part, i in zip(X, list(range(len(X))))`